Refactored testoutput helper usage
