+++ Workshops/EventDrivenArchitecture/03-BusinessProcesses/Version1-Aggregates/GroupCheckouts/GroupCheckoutEvent.cs
