+++ Workshops/EventDrivenArchitecture/03-BusinessProcesses/Version2-ImmutableEntities/GroupCheckouts/GroupCheckoutEvents.cs
@@ -1,0 +1,42 @@
+using BusinessProcesses.Version2_ImmutableEntities.Core;
+
+namespace BusinessProcesses.Version2_ImmutableEntities.GroupCheckouts;
+
+using static GroupCheckoutEvent;
+
+public abstract record GroupCheckoutEvent
+{
+    public record GroupCheckoutInitiated(
+        Guid GroupCheckoutId,
+        Guid ClerkId,
+        Guid[] GuestStayIds,
+        DateTimeOffset InitiatedAt
+    ): GroupCheckoutEvent;
+
+    public record GuestCheckoutCompleted(
+        Guid GroupCheckoutId,
+        Guid GuestStayId,
+        DateTimeOffset CompletedAt
+    ): GroupCheckoutEvent;
+
+    public record GuestCheckoutFailed(
+        Guid GroupCheckoutId,
+        Guid GuestStayId,
+        DateTimeOffset FailedAt
+    ): GroupCheckoutEvent;
+
+    public record GroupCheckoutCompleted(
+        Guid GroupCheckoutId,
+        Guid[] CompletedCheckouts,
+        DateTimeOffset CompletedAt
+    ): GroupCheckoutEvent;
+
+    public record GroupCheckoutFailed(
+        Guid GroupCheckoutId,
+        Guid[] CompletedCheckouts,
+        Guid[] FailedCheckouts,
+        DateTimeOffset FailedAt
+    ): GroupCheckoutEvent;
+
+    private GroupCheckoutEvent() { }
+}
