+++ Workshops/EventDrivenArchitecture/EventDrivenArchitecture.sln
@@ -20,6 +20,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "01-EventsDefinition", "Solu
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "02-EntitiesDefinition", "Solutions\02-EntitiesDefinition\02-EntitiesDefinition.csproj", "{4901FBDE-1248-402F-89F2-8027655B703D}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "03-BusinessProcesses", "Solutions\03-BusinessProcesses\03-BusinessProcesses.csproj", "{0C7E9C5A-090E-4D48-A2CC-D60AF87591AA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -62,9 +64,14 @@ Global
 		{4901FBDE-1248-402F-89F2-8027655B703D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4901FBDE-1248-402F-89F2-8027655B703D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4901FBDE-1248-402F-89F2-8027655B703D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0C7E9C5A-090E-4D48-A2CC-D60AF87591AA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0C7E9C5A-090E-4D48-A2CC-D60AF87591AA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0C7E9C5A-090E-4D48-A2CC-D60AF87591AA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0C7E9C5A-090E-4D48-A2CC-D60AF87591AA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{7312A099-6FAD-462A-8D94-FB817B531A27} = {96700308-E324-4D7A-854D-9ED17ECEE9B7}
 		{4901FBDE-1248-402F-89F2-8027655B703D} = {96700308-E324-4D7A-854D-9ED17ECEE9B7}
+		{0C7E9C5A-090E-4D48-A2CC-D60AF87591AA} = {96700308-E324-4D7A-854D-9ED17ECEE9B7}
 	EndGlobalSection
 EndGlobal
