+++ Workshops/EventDrivenArchitecture/Solutions/02-EntitiesDefinition/Solution2-ImmutableEntities/EntityDefinitionTests.cs
@@ -14,8 +14,6 @@ using static GroupCheckoutCommand;
 
 public class EntityDefinitionTests
 {
-    private readonly ITestOutputHelper testOutputHelper;
-
     [Fact]
     public async Task CheckingInGuest_Succeeds()
     {
@@ -163,6 +161,7 @@ public class EntityDefinitionTests
     private readonly GuestStayFacade guestStayFacade;
     private readonly Faker generate = new();
     private readonly DateTimeOffset now = DateTimeOffset.Now;
+    private readonly ITestOutputHelper testOutputHelper;
 
     public EntityDefinitionTests(ITestOutputHelper testOutputHelper)
     {
