+++ Workshops/EventDrivenArchitecture/Solutions/03-BusinessProcesses/UnitTest1.cs
@@ -1,0 +1,11 @@
+using Xunit;
+
+namespace BusinessProcesses;
+
+public class UnitTest1
+{
+    [Fact]
+    public void Test1()
+    {
+    }
+}
